pyproject.toml
[project]
name = "p1-health-agent"
version = "1.0.0"
description = "AI Health Insight Agent - ML-powered health data analysis, risk scoring, and personalized recommendations"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.6.0",
    "plotly>=5.24.0",
    "fpdf2>=2.8.0",
    "numpy>=2.0.0",
]

[project.scripts]
health-agent = "src.app:main"
